Adapted pipeline to work w/ new getfiles script, implemented multi_photom.py, allowing photometry to be run on all chips w/ 1 command.

+++ photomitrus/multi_master.py
@@ -6,6 +6,7 @@ from subprocess import Popen
 import shutil
 
 from photomitrus.getdata import download_data
+from photomitrus.getfiles import get_data_files
 from photomitrus.master import master
 from photomitrus.settings import PIPELINE_DEFAULT_DIR
 
@@ -25,6 +26,29 @@ def parentcreation(target, date, band):
     return parent_dir, field_dir
 
 #%%
+
+
+def datalistdownload(parentdir, target, band, date):
+    print('Verifying parent dir: %s' % parentdir)
+    if not os.path.isdir(parentdir):
+        os.mkdir(parentdir)
+    else:
+        pass
+    if band == 'Z':
+        try:
+            full_ramp_list, m_list = get_data_files(date=date, objname=target, filter1=band, filter2='Open')
+            if any(lst for lst in m_list):
+                print('Missing files!: ', m_list)
+        except:
+            print('Error fetching data!')
+    else:
+        try:
+            full_ramp_list, m_list = get_data_files(date=date, objname=target, filter1='Open', filter2=band)
+            if any(lst for lst in m_list):
+                print('Missing files!: ', m_list)
+        except:
+            print('Error fetching data!')
+    return full_ramp_list, m_list
 
 
 def datadownload(parentdir, target, band, date, chip):
@@ -49,7 +73,7 @@ def datadownload(parentdir, target, band, date, chip):
 #%%
 
 
-def refineprocess(parentdir, chip, band, rot_val=48, sky_override_path=None, removal=False):
+def refineprocess(parentdir, chip, band, rot_val=48, sky_override_path=None, removal=False, fullramplist=None):
     if chip == 1 or chip == 2:
         sigma = 4
     elif chip == 3 or chip == 4:
@@ -57,10 +81,10 @@ def refineprocess(parentdir, chip, band, rot_val=48, sky_override_path=None, rem
     else:
         sigma = None
     master(parentdir=parentdir, chip=chip, band=band, sigma=sigma, rot_val=rot_val,net_refine=True,
-           sky_override=sky_override_path, removal=removal)
+           sky_override=sky_override_path, removal=removal, fullramplist=fullramplist)
 
 
-def shiftprocess(parentdir, chip, band, rot_val=48, sky_override_path=None, removal=False):
+def shiftprocess(parentdir, chip, band, rot_val=48, sky_override_path=None, removal=False, fullramplist=None):
     if chip == 1 or chip == 2:
         sigma = 4
     elif chip == 3 or chip == 4:
@@ -68,10 +92,10 @@ def shiftprocess(parentdir, chip, band, rot_val=48, sky_override_path=None, remo
     else:
         sigma = None
     master(parentdir=parentdir, chip=chip, band=band, sigma=sigma, rot_val=rot_val, sky_override=sky_override_path,
-           removal=removal)
+           removal=removal, fullramplist=fullramplist)
 
 
-def baseprocess(parentdir, chip, band, rot_val=48, sky_override_path=None, removal=False):
+def baseprocess(parentdir, chip, band, rot_val=48, sky_override_path=None, removal=False, fullramplist=None):
     if chip == 1 or chip == 2:
         sigma = 4
     elif chip == 3 or chip == 4:
@@ -79,7 +103,7 @@ def baseprocess(parentdir, chip, band, rot_val=48, sky_override_path=None, remov
     else:
         sigma = None
     master(parentdir=parentdir, chip=chip, band=band, sigma=sigma, rot_val=rot_val, no_shift=True,
-           sky_override=sky_override_path, removal=removal)
+           sky_override=sky_override_path, removal=removal, fullramplist=fullramplist)
 
 
 def processparallel(target, date, band, chips):
@@ -102,7 +126,7 @@ def processparallel(target, date, band, chips):
 
 def multi_master(
         target, date, band, chip=None, parentdir=None, rot_val=48, no_shift=False, astromnet=False, parallel=False,
-        no_download=False, sky_override_path=None, removal=False
+        no_download=False, sky_override_path=None, removal=False, download_files=False
 ):
 
     if parentdir:
@@ -116,33 +140,30 @@ def multi_master(
         chips = chip.split(',')
         chips = [int(f) for f in chips]
 
-    if no_download:
-        pass
-    else:
-        chips_str = ','.join(str(x) for x in chips)
-        datadownload(chosen_parent, target, band, date, chips_str)
-
-    if chip:
-        chip_path = os.path.join(chosen_parent, 'C%s/' % chips[0])
-        print(chip_path)
-    else:
-        chip_path = os.path.join(chosen_parent, 'C1/')
-
-    if not os.listdir(chip_path):
-        print('Error downloading, No data! Or perhaps wrong date or target?')
-        if parentdir:
+    if download_files:
+        if no_download:
             pass
         else:
-            print('Removing default directory: %s' % field_dir)
-            try:
-                os.chdir(PIPELINE_DEFAULT_DIR)
-                shutil.rmtree(field_dir, ignore_errors=True)
-            except FileNotFoundError:
-                print('Directory already no longer exists.')
-    elif parallel:
-        print('Processing all chosen chips in parallel!')
-        processparallel(target, date, band, chips)
-    else:
+            chips_str = ','.join(str(x) for x in chips)
+            datadownload(chosen_parent, target, band, date, chips_str)
+
+        if chip:
+            chip_path = os.path.join(chosen_parent, 'C%s/' % chips[0])
+            print(chip_path)
+        else:
+            chip_path = os.path.join(chosen_parent, 'C1/')
+
+        if not os.listdir(chip_path):
+            print('Error downloading, No data! Or perhaps wrong date or target?')
+            if parentdir:
+                pass
+            else:
+                print('Removing default directory: %s' % field_dir)
+                try:
+                    os.chdir(PIPELINE_DEFAULT_DIR)
+                    shutil.rmtree(field_dir, ignore_errors=True)
+                except FileNotFoundError:
+                    print('Directory already no longer exists.')
         for f in chips:
             if astromnet:
                 refineprocess(chosen_parent, f, band, rot_val, sky_override_path, removal=removal)
@@ -150,13 +171,43 @@ def multi_master(
                 shiftprocess(chosen_parent, f, band, rot_val, sky_override_path, removal=removal)
             else:
                 baseprocess(chosen_parent, f, band, rot_val, sky_override_path, removal=removal)
+    else:
+        full_ramp_list, m_list = datalistdownload(chosen_parent, target, band, date)
+        if all(not lst for lst in full_ramp_list):
+            print('Error finding files, No data! Or perhaps wrong date or target?')
+            if parentdir:
+                pass
+            else:
+                print('Removing default directory: %s' % field_dir)
+                try:
+                    os.chdir(PIPELINE_DEFAULT_DIR)
+                    shutil.rmtree(field_dir, ignore_errors=True)
+                except FileNotFoundError:
+                    print('Directory already no longer exists.')
+        # if parallel:
+        #     print('Processing all chosen chips in parallel!')
+        #     processparallel(target, date, band, chips)
+        for f in chips:
+            if astromnet:
+                refineprocess(chosen_parent, f, band, rot_val, sky_override_path, removal=removal,
+                              fullramplist=full_ramp_list)
+            elif not no_shift:
+                shiftprocess(chosen_parent, f, band, rot_val, sky_override_path, removal=removal,
+                             fullramplist=full_ramp_list)
+            else:
+                baseprocess(chosen_parent, f, band, rot_val, sky_override_path, removal=removal,
+                            fullramplist=full_ramp_list)
 
 
 def main():
     parser = argparse.ArgumentParser(description='Use to process whole observations (all chips)')
-    parser.add_argument('-parallel', action='store_true', help='optional flag, process multiple chips simultaneously,'
-                                                               ' only use on obs. w/ small amount of images!')
-    parser.add_argument('-no_download', action='store_true', help='optional flag, use if you already have the data')
+    # parser.add_argument('-parallel', action='store_true', help='optional flag, process multiple chips simultaneously,'
+    #                                                            ' only use on obs. w/ small amount of images!')
+    parser.add_argument('-download_files', action='store_true', help='optional flag, use if you want to download '
+                                                                     'through old method (scp), new method passes file '
+                                                                     'paths (saves space and time)')
+    parser.add_argument('-no_download', action='store_true', help='optional flag, use if you *ALREADY* have the data'
+                                                                  'downloaded, *NOT* to use new file path method')
     parser.add_argument('-no_shift', action='store_true', help='optional flag, DO NOT use astrometric shift'
                                                                ' script in place of astrom.net, will not use either (shift is default)')
     parser.add_argument('-astromnet', action='store_true', help='optional flag, use astrom.net to reinforce astrometry')
@@ -169,7 +220,8 @@ def main():
     parser.add_argument('-target', type=str, help='[str] target field, objname in log, ex. "field1234"')
     parser.add_argument('-date', type=str, help='[str] date of observation, in yyyymmdd format')
     parser.add_argument('-band', type=str, help='[str] filter, ex. "J"')
-    parser.add_argument('-chip', type=str, help='[str] Optional, use to process specific chips',default=None)
+    parser.add_argument('-chip', type=str, help='[str] Optional, use to process specific chips, use "1,2,3,4"'
+                                                ' format.',default=None)
     parser.add_argument('-rot_val', type=float, help='[float] optional, put in your rot angle in deg,'
                                                      ' if you had a non-default rotation angle in your obs'
                                                      ' (default = 48 deg or 172800")', default=48)
@@ -179,7 +231,7 @@ def main():
     args, unknown = parser.parse_known_args()
 
     multi_master(args.target, args.date, args.band, args.chip, args.parent, args.rot_val, args.no_shift, args.astromnet,
-                 args.parallel, args.no_download, args.sky_override, args.removal)
+                 args.no_download, args.sky_override, args.removal, args.download_files)
 
 
 if __name__ == "__main__":
